gfx fix

+++ src/device/ili9486.h
@@ -19,10 +19,10 @@ public:
     void init(Orientation ori);
     void setOrientation(Orientation ori);
     
-	void fillRect(int x, int y, int width, int height, uint16_t color);
-    void copyRect(int x, int y, int width, int height, const uint16_t *buffer);
+	void fillRect(int x, int y, int width, int height, uint32_t color);
+    void copyRect(int x, int y, int width, int height, const uint8_t *buffer);
     
-    virtual void setPixel(int x, int y, uint16_t color);
+    virtual void setPixel(int x, int y, uint32_t color);
 //    virtual uint16_t pixel(int x, int y); // THIS DISPLAY IMPLEMENTATION IS NOT READABLE!!!
     
 private:
